requirements.txt
streamlit
streamlit-option-menu
pandas
plotly
Pillow
Babel